Revert "- Switching to Qt6 support"
This reverts commit cb8571c7540deb4247497e33d330c72f1ba83a3a.

+++ src/QPlot/QPlotButton.cpp
@@ -16,7 +16,7 @@ Button::Button(QCustomPlot *parentPlot,
   setPixmap(pixmap);
   setSelectable(false);
 
-  if (second_point == static_cast<int>(Qt::AlignBottom | Qt::AlignRight))
+  if (second_point == (Qt::AlignBottom | Qt::AlignRight))
   {
     bottomRight->setType(QCPItemPosition::ptAbsolute);
     bottomRight->setParentAnchor(topLeft);
